+++ superduper/backends/mongodb/query.py
@@ -401,33 +401,8 @@ class MongoQuery(Query):
     @property
     @applies_to('insert_many', 'insert_one')
     def documents(self):
-        """Return the documents to insert."""
-
-        def _wrap_document(document):
-            if not isinstance(document, Document):
-                if isinstance(document, dict):
-                    document = Document(document)
-                else:
-                    document = Document({'_base': document})
-            return document
-
-        def _update_part(documents):
-            nonlocal self
-            doc_args = (documents, *self.parts[0][1][1:])
-            insert_part = (self.parts[0][0], doc_args, self.parts[0][2])
-            return [insert_part] + self.parts[1:]
-
-        if self.parts[0][0] == 'insert_many':
-            documents = self.parts[0][1][0]
-            wrapped_documents = []
-            for document in documents:
-                document = _wrap_document(document)
-                wrapped_documents.append(document)
-            self.parts = _update_part(wrapped_documents)
-            return wrapped_documents
-        document = _wrap_document(self.parts[0][1][0])
-        self.parts = _update_part(document)
-        return [document]
+        """Return the documents."""
+        return super().documents
 
     @property
     def primary_id(self):