+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Modify the field name output to _outputs.predict_id in the model results of Ibis.
 - Remove the document_embed function.
 - Support MongoDB outputs query
+- Make "create a table" compulsory
 
 #### Bug Fixes